All commands now moved out of main. Most other functions moved as well.
Several new cogs created, massive re-org of command and method location,
lots of pep8 cleanup
