+++ kyogre/exts/raidparty.py
@@ -2,13 +2,12 @@ import asyncio
 import copy
 import datetime
 import re
-import textwrap
 import time
 
 import discord
 from discord.ext import commands
 
-from kyogre import checks, list_helpers, raid_helpers, raid_lobby_helpers
+from kyogre import checks, embed_utils, utils
 from kyogre.exts.pokemon import Pokemon
 
 
@@ -147,7 +146,8 @@ class RaidParty(commands.Cog):
         if isinstance(result, list):
             count = result[0]
             partylist = result[1]
-            await list_helpers._maybe(ctx, self.bot, count, partylist, eta, entered_interest)
+            listmgmt_cog = self.bot.cogs.get('ListManagement')
+            await listmgmt_cog.maybe(ctx, self.bot, count, partylist, eta, entered_interest)
 
     @commands.command(aliases=['c'])
     @checks.activechannel()
@@ -168,7 +168,8 @@ class RaidParty(commands.Cog):
         if isinstance(result, list):
             count = result[0]
             partylist = result[1]
-            await list_helpers._coming(ctx, self.bot, count, partylist, eta, entered_interest)
+            listmgmt_cog = self.bot.cogs.get('ListManagement')
+            await listmgmt_cog.coming(ctx, self.bot, count, partylist, eta, entered_interest)
 
     @commands.command(aliases=['h'])
     @checks.activechannel()
@@ -188,7 +189,8 @@ class RaidParty(commands.Cog):
         if isinstance(result, list):
             count = result[0]
             partylist = result[1]
-            await list_helpers._here(ctx, self.bot, count, partylist, entered_interest)
+            listmgmt_cog = self.bot.cogs.get('ListManagement')
+            await listmgmt_cog.here(ctx, self.bot, count, partylist, entered_interest)
 
     async def _party_status(self, ctx, total, teamcounts):
         channel = ctx.channel
@@ -315,7 +317,8 @@ class RaidParty(commands.Cog):
         else:
             await channel.send('{member} is entering the lobby with a total of {trainer_count} trainers!'
                                .format(member=trainer.mention, trainer_count=count))
-            regions = raid_helpers.get_channel_regions(channel, 'raid', self.bot.guild_dict)
+            utils_cog = self.bot.cogs.get('Utilities')
+            regions = utils_cog.get_channel_regions(channel, 'raid')
             joined = self.bot.guild_dict[guild.id].setdefault('trainers', {})\
                          .setdefault(regions[0], {})\
                          .setdefault(trainer.id, {})\
@@ -328,8 +331,8 @@ class RaidParty(commands.Cog):
         self.bot.guild_dict[guild.id]['raidchannel_dict'][channel.id]['trainer_dict'] = trainer_dict
         regions = self.bot.guild_dict[guild.id]['raidchannel_dict'][channel.id].get('regions', None)
         if regions:
-            await list_helpers.update_listing_channels(self.bot, channel.guild,
-                                                       'raid', edit=True, regions=regions)
+            listmgmt_cog = self.bot.cogs.get('ListManagement')
+            await listmgmt_cog.update_listing_channels(channel.guild, 'raid', edit=True, regions=regions)
 
     @commands.command(aliases=['x'])
     @checks.raidchannel()
@@ -339,7 +342,8 @@ class RaidParty(commands.Cog):
         **Usage**: `!cancel/x`
         Removes you and your party from the list of trainers who are "coming" or "here".
         Or removes you and your party from the active lobby."""
-        await list_helpers._cancel(ctx, self.bot)
+        listmgmt_cog = self.bot.cogs.get('ListManagement')
+        await listmgmt_cog.cancel(ctx, self.bot)
 
     @commands.command(aliases=['s'])
     @checks.activeraidchannel()
@@ -350,7 +354,127 @@ class RaidParty(commands.Cog):
         Sends a message notifying all trainers who are at the raid and clears the waiting list.
         Starts a 2 minute lobby countdown during which time trainers can join this lobby using `!lobby`.
         Users who are waiting for a second group must reannounce with `!here`."""
-        await raid_lobby_helpers._starting(ctx, self.bot, team)
+        guild_dict = self.bot.guild_dict
+        channel = ctx.channel
+        guild = ctx.guild
+        ctx_startinglist = []
+        team_list = []
+        ctx.team_names = ["mystic", "valor", "instinct", "unknown"]
+        team = team if team and team.lower() in ctx.team_names else "all"
+        ctx.trainer_dict = copy.deepcopy(guild_dict[guild.id]['raidchannel_dict'][channel.id]['trainer_dict'])
+        regions = guild_dict[guild.id]['raidchannel_dict'][channel.id]['regions']
+        if guild_dict[guild.id]['raidchannel_dict'][channel.id].get('type', None) == 'egg':
+            if guild_dict[guild.id]['raidchannel_dict'][channel.id]['exp'] - 60 < datetime.datetime.now().timestamp():
+                starting_str = "Please tell me which raid boss has hatched before starting your lobby."
+            else:
+                starting_str = "How can you start when the egg hasn't hatched!?"
+            await channel.send(starting_str)
+            return
+        if guild_dict[guild.id]['raidchannel_dict'][channel.id].get('lobby', False):
+            starting_str = "Please wait for the group in the lobby to enter the raid."
+            await channel.send(starting_str)
+            return
+        trainer_joined = False
+        for trainer in ctx.trainer_dict:
+            count = ctx.trainer_dict[trainer]['count']
+            user = guild.get_member(trainer)
+            if team in ctx.team_names:
+                if ctx.trainer_dict[trainer]['party'][team]:
+                    team_list.append(user.id)
+                teamcount = ctx.trainer_dict[trainer]['party'][team]
+                herecount = ctx.trainer_dict[trainer]['status']['here']
+                lobbycount = ctx.trainer_dict[trainer]['status']['lobby']
+                if ctx.trainer_dict[trainer]['status']['here'] and (user.id in team_list):
+                    ctx.trainer_dict[trainer]['status'] = {'maybe': 0, 'coming': 0, 'here': herecount - teamcount,
+                                                           'lobby': lobbycount + teamcount}
+                    trainer_joined = True
+                    ctx_startinglist.append(user.mention)
+            else:
+                if ctx.trainer_dict[trainer]['status']['here'] and (user.id in team_list or team == "all"):
+                    ctx.trainer_dict[trainer]['status'] = {'maybe': 0, 'coming': 0, 'here': 0, 'lobby': count}
+                    trainer_joined = True
+                    ctx_startinglist.append(user.mention)
+            if trainer_joined:
+                joined = guild_dict[guild.id].setdefault('trainers', {}).setdefault(regions[0], {}).setdefault(trainer,
+                                                                                                               {}).setdefault(
+                    'joined', 0) + 1
+                guild_dict[guild.id]['trainers'][regions[0]][trainer]['joined'] = joined
+
+        if len(ctx_startinglist) == 0:
+            starting_str = "How can you start when there's no one waiting at this raid!?"
+            await channel.send(starting_str)
+            return
+        guild_dict[guild.id]['raidchannel_dict'][channel.id]['trainer_dict'] = ctx.trainer_dict
+        starttime = guild_dict[guild.id]['raidchannel_dict'][channel.id].get('starttime', None)
+        if starttime:
+            timestr = ' to start at **{}** '.format(starttime.strftime('%I:%M %p (%H:%M)'))
+            guild_dict[guild.id]['raidchannel_dict'][channel.id]['starttime'] = None
+        else:
+            timestr = ' '
+        starting_str = 'Starting - The group that was waiting{timestr}is starting the raid! Trainers {trainer_list}, if you are not in this group and are waiting for the next group, please respond with {here_emoji} or **!here**. If you need to ask those that just started to back out of their lobby, use **!backout**'.format(
+            timestr=timestr, trainer_list=', '.join(ctx_startinglist),
+            here_emoji=utils.parse_emoji(guild, self.bot.config['here_id']))
+        guild_dict[guild.id]['raidchannel_dict'][channel.id]['lobby'] = {"exp": time.time() + 120, "team": team}
+        if starttime:
+            starting_str += '\n\nThe start time has also been cleared, new groups can set a new start time wtih **!starttime HH:MM AM/PM** (You can also omit AM/PM and use 24-hour time!).'
+            report_channel = self.bot.get_channel(guild_dict[guild.id]['raidchannel_dict'][channel.id]['reportcity'])
+            raidmsg = await channel.fetch_message(guild_dict[guild.id]['raidchannel_dict'][channel.id]['raidmessage'])
+            reportmsg = await report_channel.fetch_message(
+                guild_dict[guild.id]['raidchannel_dict'][channel.id]['raidreport'])
+            embed = raidmsg.embeds[0]
+            embed_indices = await embed_utils.get_embed_field_indices(embed)
+            embed.set_field_at(embed_indices["next"], name="**Next Group**", value="Set with **!starttime**",
+                               inline=True)
+            try:
+                await raidmsg.edit(content=raidmsg.content, embed=embed)
+            except discord.errors.NotFound:
+                pass
+            try:
+                await reportmsg.edit(content=reportmsg.content, embed=embed)
+            except discord.errors.NotFound:
+                pass
+        await channel.send(starting_str)
+        ctx.bot.loop.create_task(self.lobby_countdown(ctx, team))
+        regions = guild_dict[channel.guild.id]['raidchannel_dict'][channel.id].get('regions', None)
+        if regions:
+            listmgmt_cog = self.bot.cogs.get('ListManagement')
+            await listmgmt_cog.update_listing_channels(channel.guild, 'raid', edit=True, regions=regions)
+
+    async def lobby_countdown(self, ctx, team):
+        guild_dict, raid_info = self.bot.guild_dict, self.bot.raid_info
+        await asyncio.sleep(120)
+        if ('lobby' not in guild_dict[ctx.guild.id]['raidchannel_dict'][ctx.channel.id]) or (
+                time.time() < guild_dict[ctx.guild.id]['raidchannel_dict'][ctx.channel.id]['lobby']['exp']):
+            return
+        ctx_lobbycount = 0
+        trainer_delete_list = []
+        for trainer in ctx.trainer_dict:
+            if ctx.trainer_dict[trainer]['status']['lobby']:
+                ctx_lobbycount += ctx.trainer_dict[trainer]['status']['lobby']
+                trainer_delete_list.append(trainer)
+        if ctx_lobbycount > 0:
+            await ctx.channel.send('The group of {count} in the lobby has entered the raid! Wish them luck!'.format(
+                count=str(ctx_lobbycount)))
+        for trainer in trainer_delete_list:
+            if team in ctx.team_names:
+                herecount = ctx.trainer_dict[trainer]['status'].get('here', 0)
+                teamcount = ctx.trainer_dict[trainer]['party'][team]
+                ctx.trainer_dict[trainer]['status'] = {'maybe': 0, 'coming': 0, 'here': herecount - teamcount,
+                                                       'lobby': ctx_lobbycount}
+                ctx.trainer_dict[trainer]['party'][team] = 0
+                ctx.trainer_dict[trainer]['count'] = ctx.trainer_dict[trainer]['count'] - teamcount
+            else:
+                del ctx.trainer_dict[trainer]
+        try:
+            del guild_dict[ctx.guild.id]['raidchannel_dict'][ctx.channel.id]['lobby']
+        except KeyError:
+            pass
+        listmgmt_cog = self.bot.cogs.get('ListManagement')
+        await listmgmt_cog.edit_party(ctx, ctx.channel, ctx.author)
+        guild_dict[ctx.guild.id]['raidchannel_dict'][ctx.channel.id]['trainer_dict'] = ctx.trainer_dict
+        regions = guild_dict[ctx.channel.guild.id]['raidchannel_dict'][ctx.channel.id].get('regions', None)
+        if regions:
+            await listmgmt_cog.update_listing_channels(ctx.guild, 'raid', edit=True, regions=regions)
 
     @commands.command()
     @checks.activeraidchannel()
@@ -360,8 +484,71 @@ class RaidParty(commands.Cog):
         **Usage**: `!backout`
         Will alert all trainers in the lobby that a backout is requested.
         Those trainers can exit the lobby with `!cancel`."""
-        await raid_lobby_helpers._backout(ctx, self.bot)
-         
+        guild_dict = self.bot.guild_dict
+        message = ctx.message
+        channel = message.channel
+        author = message.author
+        guild = channel.guild
+        trainer_dict = guild_dict[guild.id]['raidchannel_dict'][channel.id]['trainer_dict']
+        if (author.id in trainer_dict) and (trainer_dict[author.id]['status']['lobby']):
+            count = trainer_dict[author.id]['count']
+            trainer_dict[author.id]['status'] = {'maybe': 0, 'coming': 0, 'here': count, 'lobby': 0}
+            lobby_list = []
+            for trainer in trainer_dict:
+                count = trainer_dict[trainer]['count']
+                if trainer_dict[trainer]['status']['lobby']:
+                    user = guild.get_member(trainer)
+                    lobby_list.append(user.mention)
+                    trainer_dict[trainer]['status'] = {'maybe': 0, 'coming': 0, 'here': count, 'lobby': 0}
+            if not lobby_list:
+                await channel.send("There's no one else in the lobby for this raid!")
+                try:
+                    del guild_dict[guild.id]['raidchannel_dict'][channel.id]['lobby']
+                except KeyError:
+                    pass
+                return
+            await channel.send('Backout - {author} has indicated that the group consisting of {lobby_list} and the '
+                               'people with them has backed out of the lobby! If this is inaccurate, please use '
+                               '**!lobby** or **!cancel** to help me keep my lists accurate!'
+                               .format(author=author.mention, lobby_list=', '.join(lobby_list)))
+            try:
+                del guild_dict[guild.id]['raidchannel_dict'][channel.id]['lobby']
+            except KeyError:
+                pass
+        else:
+            lobby_list = []
+            trainer_list = []
+            for trainer in trainer_dict:
+                if trainer_dict[trainer]['status']['lobby']:
+                    user = guild.get_member(trainer)
+                    lobby_list.append(user.mention)
+                    trainer_list.append(trainer)
+            if (not lobby_list):
+                await channel.send("There's no one in the lobby for this raid!")
+                return
+
+            backoutmsg = await channel.send(
+                'Backout - {author} has requested a backout! If one of the following trainers reacts with the '
+                'check mark, I will assume the group is backing out of the raid lobby as requested! {lobby_list}'
+                    .format(author=author.mention, lobby_list=', '.join(lobby_list)))
+            try:
+                timeout = False
+                res, reactuser = await utils.simple_ask(self.bot, backoutmsg, channel, trainer_list, react_list=['✅'])
+            except TypeError:
+                timeout = True
+            if not timeout and res.emoji == '✅':
+                for trainer in trainer_list:
+                    count = trainer_dict[trainer]['count']
+                    if trainer in trainer_dict:
+                        trainer_dict[trainer]['status'] = {'maybe': 0, 'coming': 0, 'here': count, 'lobby': 0}
+                await channel.send('{user} confirmed the group is backing out!'.format(user=reactuser.mention))
+                try:
+                    del guild_dict[guild.id]['raidchannel_dict'][channel.id]['lobby']
+                except KeyError:
+                    pass
+            else:
+                return
+
 
 def setup(bot):
     bot.add_cog(RaidParty(bot))
